Sum of Digits Python
In this challenge you have to take input 3 numbers. Multiply first two numbers and then add the result with the third number. Then the we have to add every digit of the result and return the answer. You have to do the same steps for every 3 digit pair you enter.